requirements.txt
attrs>=19.1,<20.0
cattrs>=0.9.0,<0.10.0
click>=7.0,<8.0
gherkin-official>=4.1,<5.0
